--- js/form-validation.js
@@ -1,1 +1,0 @@
-
